requirements.txt
aws-psycopg2==1.2.1
boto3==1.18.1
botocore==1.21.1
certifi==2021.5.30
charset-normalizer==2.0.3
Django==2.2.24
django-cors-headers==3.7.0
django-storages==1.11.1
djangorestframework==3.12.4
idna==3.2
jmespath==0.10.0
python-dateutil==2.8.2
python-http-client==3.3.2
pytz==2021.1
requests==2.26.0
s3transfer==0.5.0
sendgrid==6.7.1
six==1.16.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.1
stripe==5.4.0
timeloop==1.0.2
urllib3==1.26.6
